+++ skytemple_files/script/ssa_sse_sss/object.py
@@ -17,6 +17,8 @@
 import logging
 import warnings
 
+from range_typed_integers import u16, u8
+
 from skytemple_files.common.ppmdu_config.script_data import Pmd2ScriptData, Pmd2ScriptObject
 from skytemple_files.common.util import AutoString
 from skytemple_files.script.ssa_sse_sss.position import SsaPosition
@@ -29,7 +31,7 @@ class SsaObject(AutoString):
             self.object = scriptdata.objects__by_id[object_id]
         except KeyError:
             logger.warning(f"Unknown object id: {object_id}")
-            self.object = Pmd2ScriptObject(object_id, 0, 0, 0, 'UNKNOWN')
+            self.object = Pmd2ScriptObject(object_id, u16(0), u16(0), u8(0), 'UNKNOWN')
         self.hitbox_w = htibox_w
         self.hitbox_h = hitbox_h
         self.pos = pos
